# services/identity/internal/httpapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Facinect Identity API",
    "description": "Auth, session, and partner membership APIs (local gateway).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Traefik gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "token_type": { "type": "string", "example": "Bearer" },
          "expires_in": { "type": "integer" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string" },
          "full_name": { "type": "string" },
          "avatar_url": { "type": "string", "nullable": true },
          "has_password": { "type": "boolean" }
        }
      },
      "Facility": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "role": { "type": "string" },
          "page_access": {
            "type": "array",
            "items": { "type": "string" },
            "nullable": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/auth/check-email": {
      "post": {
        "tags": ["auth"],
        "summary": "Check if email exists",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": { "type": "boolean" },
                    "email": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Email + password login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@facinect.local"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "ChangeMe_Admin_123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens (+ session fields may be included)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenPair" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/google/start": {
      "get": {
        "tags": ["google"],
        "summary": "Start Google OAuth (PKCE)",
        "parameters": [
          {
            "name": "redirect",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Optional post-login path"
          }
        ],
        "responses": {
          "302": { "description": "Redirect to Google" }
        }
      }
    },
    "/v1/auth/google/callback": {
      "get": {
        "tags": ["google"],
        "summary": "Google OAuth callback",
        "parameters": [
          { "name": "code", "in": "query", "schema": { "type": "string" } },
          { "name": "state", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "302": { "description": "Redirect to /login?handoff=..." }
        }
      }
    },
    "/v1/auth/google/handoff": {
      "post": {
        "tags": ["google"],
        "summary": "Exchange handoff code for tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["handoff"],
                "properties": {
                  "handoff": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenPair" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/otp/request": {
      "post": {
        "tags": ["otp"],
        "summary": "Request OTP (local may return dev_otp)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string" },
                  "whatsappNo": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP requested" },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "tags": ["otp"],
        "summary": "Verify OTP → tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string" },
                  "whatsappNo": { "type": "string" },
                  "code": { "type": "string" },
                  "otp": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenPair" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/token/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Rotate refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenPair" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Revoke refresh token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/password": {
      "post": {
        "tags": ["auth"],
        "summary": "Set or update password (no current password required)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "new_password": { "type": "string", "minLength": 8 },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": ["session"],
        "summary": "Current user + facilities from JWT context",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "facilities": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Facility" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/session": {
      "get": {
        "tags": ["session"],
        "summary": "Live membership refresh from DB",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Fresh session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "facilities": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Facility" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "No facilities assigned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/v1/auth/session/refresh": {
      "post": {
        "tags": ["session"],
        "summary": "Same as GET /session (POST alias)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Fresh session" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "No facilities assigned" }
        }
      }
    },
...
  }
}